+++ packages/ui/components/common/Icon/GlobeIcon.tsx
@@ -1,0 +1,5 @@
+import GlobeSvg from "../../../../../root/assets/icon/globe.svg";
+
+export default function GlobeIcon({ className }: { className?: string }) {
+  return <GlobeSvg className={className} />;
+}