+++ packages/ui/components/common/Icon/IcArrowRight16.tsx
@@ -1,0 +1,5 @@
+import IcArrowRight16Svg from "../../../../../root/assets/icon/ic-arrow-right-16.svg";
+
+export function IcArrowRight16({ className }: { className?: string }) {
+  return <IcArrowRight16Svg className={className} />;
+}