+++ packages/ui/components/common/Icon/IcRequiredDot.tsx
@@ -1,0 +1,5 @@
+import IcRequiredDotSvg from "../../../../../root/assets/icon/ic-required-dot.svg";
+
+export function IcRequiredDot({ className }: { className?: string }) {
+  return <IcRequiredDotSvg className={className} />;
+}